refactor(generate-assets): Flatten directory structure
Feels better this way.
