+++ server/routes/components/GlobalStyles.tsx
